working on auth api
