--- bin/config/MysqlORM.dart
@@ -1,23 +1,0 @@
-// var TABLE_NAME = "";
-// var TABLE_NAME = "";
-// var TABLE_NAME = "";
-// var TABLE_NAME = "";
-
-
-
-
-// import 'MysqlConnectionController.dart';
-
-// class MysqlORM {
-//   final MysqlConnectionController _mysql = MysqlConnectionController();
-
-//   insert(table, columns, values) {
-//     // generate sql
-//   }
-
-//   delete() {}
-
-//   update() {}
-
-//   select(){}
-// }
