--- bin/config/TableUsers.dart
@@ -1,5 +1,0 @@
-// class TableUsers {
-//   var tableName = 'users';
-
-//   insert() {}
-// // }
