--- lib/dart_server.dart
@@ -1,3 +1,0 @@
-int calculate() {
-  return 6 * 7;
-}
