+++ lib/src/auth_api.dart
@@ -1,0 +1,8 @@
+import 'package:shelf_router/shelf_router.dart';
+
+class AuthAPi {
+  String? secret;
+  AuthAPi({this.secret});
+
+  Router get app {}
+}
